Ensure tests are run in CI
The test helper has functions to require cgroup tools to be installed,
but was halting `0` if undetected. This caused `mix test` to appear
to be successful even though no tests were actually run.

This changes the halt to `1` on those conditions so tests actually fail
when the linux (CI) setup is not correct